Added Angular proxy configuration to acccess backend with port 4200 during development
Previously, KHI expects the frontend dev server is running on the port 4200 and its backend is running on the port 8080.
But this is a bit hard for the remote development environment under proxied environment because it needs to proxy 2 different ports.
I changed this setup to use Angular proxy instead for development purpose. Now the Angular dev server proxies the requests under /api to the localhost:8080,
the frontend is no longer need to access localhost:8080 directly from the frontend with the special configuration only for the development.