fix bug to keep old form parameter values on new inspection dialog
The parameter form view model wasn't refreshed when user go back to the previous stepper page on the new inspection dialog. This makes the dialog show the old parameter list built for the previous inspection settings.

This change makes the view model be reset on the timing of the stepper page open. And added a progress bar for getting the form parameter response from server shown when the view model is null.