Merge main into fix/issue-274 (#284)
* Add new task types replacing legacyparser eventually (#263)

* fix test failure from #263 (#283)

Our GitHub branch protection policy didn't enforce "Require branches to be up to date before merging" under "Require status checks to pass" and it causes race condition that possibly allows PR being merged even with test failure.

* fix bug to keep old form parameter values on new inspection dialog (#282)

The parameter form view model wasn't refreshed when user go back to the previous stepper page on the new inspection dialog. This makes the dialog show the old parameter list built for the previous inspection settings.

This change makes the view model be reset on the timing of the stepper page open. And added a progress bar for getting the form parameter response from server shown when the view model is null.

* fix bug #267:  manifest generator to be aware of uid change (#272)

* fix bug #266: Fixed the grouper to check the manifest type for subresource response & request

Modified resource timeline is not simply determined from the method name or resource name in the log.
A patch against a subresource may contain its parent manifest in its response.

* Use the default string for the default value of prevRevisionBody and prevResourceUID