Added fallbacks when the fields used for main message is not available
Container log parsers and several parser tasks are depending on the main message field set reader,
but it can show noisy warnings when none of the target log fields are available.

This change fixes the issue by adding fallback fields.
Now it uses fields as a main message in the following order:
 `textPayload` > `jsonPayload.****` (**** would be `message`, `msg`...etc) > jsonPayload > labels