Adding new task based types to gather info from multiple optional tasks
This change adds NewRelationshipMergerTask and NewRelationshipDiscoveryTask.
In many inspection scenarios, it's necessary to associate information across different log sources.
For example, a log might contain an IP address, while another log maps that IP to a specific VM or container name.
However, the availability of these log sources is not always guaranteed, and consumers of this relational
data should not need to be aware of the specific tasks that provide it.

These new task types are for defining these functionality regardless of being aware of the user from provider side and the provider from user side.